fix_vagrant: workaround due to the behavior change in download_url
https://github.com/flatcar-linux/scripts/pull/8 tried to fix the
long-running issue with URLs with vagrant json files, but didn't
quite fix that. Instead it changed the output URL a little bit.
As a result, `fix_vagrant.sh` cannot fix the vagrant json files
any more. So we need to change the sed expression to make it work.

When https://github.com/flatcar-linux/scripts/pull/10 got merged,
this workaround will not be needed any more.