update-repos: do not push tags to manifest repos
We have been pushing upstream tags automatically to manifest repos of
flatcar-linux. As a result, we needed to delete remote tags for every
release. Actually we don't need to push tags during the update-repos
phase, because afterwards we will anyway push the tags manually after
git-sync has finished.

To solve this issue, split REPOS into 2 parts, a normal REPOS and
REPOS_MANIFEST. Deal with different cases with an additional parameter.