Add create-manifest helper
Currently the manifest is generated by
git-sync after release branches were created.
If a build fix is needed the manifest needs to
be updated by hand because it uses commit IDs.
Also creating point releases for an old release
is not easy currently because the branch
needs to be copied to add changes and then either
git-sync would separate release branches or the
manifest needs to be created by hand.

Add a script to create manifest entries in
a flexible way, not tied to branch creation.
The manifest uses no commit IDs but branch names
so that no new manifest is needed when build fixes
are done and a Jenkins job restarted.
The script makes point releases easy by copying
the branches via the mirror-repos-branch script,
then applying the changes and creating a new manifest.
References for scripts, coreos-overlay, and portage-stable
can be specified independently so that we can start Jenkins
builds for PRs (this requires a public manifest tag, so in the
future to avoid this we need to tell Jenkins to check out
PR branches itself after the manifest was used).