create-manifest: Expand note in manifest-template.xml.envsubst
Usually the manifest would describe which revision is used for each
repository. This is not exactly the case for Container Linux because
the ebuild files have the last say.
Call this out in the note so that people reading the manifest file
or using it to check repos out are aware of this.