cherry-pick-for: Add helper script to pick and push
Checking out a branch, ensuring that it is on the latest state,
picking a list of commits, and finally pushing is a error-prone
laborous process which this helper tries to simplify.
Only one branch is accepted and it's easy to rerun the script for
another branch. In case of failure manual resolution is required
and which commits are left can be seen from the log.