mirror-repos-branch: Limit default repos
Release branches are only needed for the three repos coreos-overlay,
portage-stable, and scripts. Release branches for the other repos
are confusing because they don't relate to the actual commit in the
ebuild at all.
Disable all other repos by default. If there should be a release branch
needed for some repo, it has to be created independent from branching the main
repos because the ebuild needs to point to the new commit in that branch.