is-pr-in-branches: Helper to see for which branches a PR was picked
Complementary to the "cherry-pick-for" tool something is needed to find
out in which branches patches are missing, so that they can be cherry
picked. Sometimes we may see that a PR is merged but it's not clear
whether the changes were picked for the other channel's branches, or
we may have picked the changes and want to verify that we didn't miss
something.
The helper tool takes a GitHub PR URL and optionally the branches to
check (default is all channel branches). It uses the patch IDs to
search for missing patches and prints them. In addition it will also
display the missing patches of the PR, but this behavior can be turned
off if it's too noisy in some cases.