cherry-pick-for: Support GitHub PRs instead of commit IDs
Often the commits to be picked originate from a GitHub PR. Currently when
this PR is merged, the commit IDs of the target branch need to be copied
as arguments for the cherry-pick-for helper.
Allow to specify the GitHub PR directly to make it simpler to use and less
error-prone because the wrong commit IDs could be copied.
Only support one PR at a time for now because this is the common case, but
if needed it's also easy to add a loop as done for the commits IDs.