coreos-overlay-diff: Print GitHub PRs for merge commits
Sometimes we need the PR link for release notes or for additional
context of some changes.
Get the PR numbers from the merge commits and print GitHub links for
them. This is a bit limited because it won't find the PRs for cherry
picks (unless a merge commit was generated for them by the
cherry-pick-for script). It also works for child repositories to find
which PRs were part of a mayday/ignition/mantle/systemd/… revision
bump.