cherry-pick-for: Cut away merging person metadata
The commit message included the "by MERGER" suffix that GitHub adds to
the HTML page title. This caused that the merge commits had a longer
message body.
Cut away the metadata about the merging person. The only difference
remaining is that the repository name is used instead of the original
branch name. Still the noise in coreos-overlay-diff is slightly reduced
by having the same PR title printed.