show-changes: add script to generate release note draft
We established the "changelog" folder to prepare the release notes
entries for all new Pull Requests.
Introduce a helper to generate the release note draft based on git
references. The script expects to run in the top level directory for
now and since we do not really use the submodules yet, it still relies
on the repository folders.