tag-release: update the submodules for the SDK container
For the SDK container we need to make sure that we update the new
coreos-overlay and portage-stable submodules in the scripts repo.
At the same time we need to remain compatible with the old way,
which is still used in LTS.
Leveraging https://github.com/flatcar-linux/scripts/pull/198 we can
use the submodule update helpers and push the update to the release
maintenance branch and also sign the commit as done before. This is
has to be done after the tags for coreos-overlay and portage-stable
are created the old way (which is not covered by the ci-automation).