upload_package: set json for Accept and Content-Type
The upload_package script is already sending json metadata with POST
requests.
However, as it does not set correct headers for that, Nebraska
considered that as `application/x-www-form-urlencoded`.
As a result, Nebraska 2.6.0 or newer fails to accept the POST requests
like:

```
{"message":"request body has an error: header Content-Type has
unexpected value \"application/x-www-form-urlencoded\""}
```

Set correct headers for the json type.