cherry-pick-for: Don't try to create branch but use existing one
The checkout was meant to either check out an existing branch or create
the local branch tracking it from the origin. The --track option
however implied to create the branch always, thus failing when it
exists.
First try to check out the existing branch, then create a local branch
as fallback action.