show-changes: Find submodules references
The usage of the same git reference for all repos only works for
releases. Dev builds may only have a tag in scripts and anyway we
would also want to support individual commits or branches as refs.
Resolve the pinned submodule refs and use them instead of requiring a
single valid ref for all three repos.