extract-initramfs-from-vmlinuz: rework to support arm64
Our arm64 vmlinuz does not have a double compression, so we need to extract
cpios after the first round already. There is also no microcode, so we need to
extract all concatenated cpios. Simplify the code by factoring out round loop
(decompress and then attempt to unpack all cpios) and rework cpio extraction by
using a while loop and a pipe with the cpio file.

Tested that this works with both amd64 and arm64.