show-changes: Allow running the script from different place
The show-changes script required to be run from the parent directory
of the scripts/coreos-overlay/portage-stable repositories. The
requirement may be inconvenient in some situations (like the repo
having a different name, or being forced to change the directory
before invoking the script). Allow pointing the show-changes script to
the locations of scripts, coreos-overlay and portage-stable
repositories. If the locations are not overridden, then the script has
the old expectations.