size-change-report: Handle the case of unchanged file listings
We generate some intermediate files using `git diff` with an option to
generate a large number of context lines in order to print all the
unchanged lines. But if both files are the same, git diff prints no
output at all. In such case, we need to copy one of the files and
prepend a space to it to have a "diff" file with unchanged lines only.

This fixes some ominous warnings about expecting a bad report.