spawn-az-vm: Add development helper to build a branch on a temporary VM
Sometimes the CI does not give enough flexiblity to iterate over fixes.
Using one's own laptop is also not always desirable.

Add a small helper script to spawn an Azure VM with "az" that can build
one or more branches, also after creation. Usage is as follows:
./spawn-az-vm --name N --ssh-key FILE --resource-group RG --location LOC --create-vm --ignition IGN BRANCHES...
The Ignition file is optional but useful to disable updates or change
the SSH port.