upload_package: Improve error handling
The get_package_id function was called with "if" which disabled set -e
and thus continued after resty returned an error. The resty function
itself doesn't want to be called with set -e and thus we have to wrap
the get_package_id call to use set -e while we can explictly react on
resty errors which disables set -e. So one time we should not use
"if"/"||" and one time we should. While at it, remove the tempfile
creation which resulted in many entries that weren't cleaned up and
which were confusing because resty also creates some and one doesn't
know which is which when debugging.