Remove submodule handling now that no release channel uses them
portage-stable and coreos-overlay used to be submodules of scripts, but
this is no longer true in any release channel, so we can drop any code
that handles the migration.

Submodule-related arguments are still passed to git where they make
sense just in case we do start using submodules that don't need special
handling.

Some of these scripts seem a little overly complex now that they only
handle the scripts repo, but let's not rewrite them now.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>