create-github-release: Add new script to make release publishing easier
This script creates a draft flatcar/scripts release in GitHub for the
given new tag. If the old tag(s) are not given, they are automatically
detected using logic from the scripts repo. The GitHub CLI (gh) is
required.

I initially tried to do this using a URL opened in the browser so that
it wouldn't require gh and you could take a look before submitting. This
didn't work because, strangely enough, GitHub does not like URLs that
are several hundred thousand characters long!

The logic from the scripts repo is a little awkward to use. Perhaps it
could be made friendlier for this script, but it's not too bad.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>