spawn-az-vm: Support building arm64 images
The default is to build amd64 images in all build commands.
Allow building arm64 images by passing the --board parameter to the
build commands. For this to work we also need qemu-user binary
translation because some package's compilation tests rely on running the
resulting binary. Set this up when creating the VM.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>